+++ tests/test_cli.py
@@ -157,7 +157,9 @@ def test_version_strategy_lifecycle(
         == '✗ curse:molinari\n  package already installed\n'
     )
     assert run('update curse:molinari').output == '✗ curse:molinari\n  package is up to date\n'
-    assert run('update curse:molinari#version_eq=80000.57-Release').output == dedent(
+    assert run(
+        'update --retain-strategies curse:molinari#version_eq=80000.57-Release'
+    ).output == dedent(
         '''\
         ✓ curse:molinari
           updated 90200.82-Release to 80000.57-Release with new strategies:
@@ -177,7 +179,7 @@ def test_version_strategy_lifecycle(
         == '✓ curse:molinari\n  installed 80000.57-Release\n'
     )
     assert run('update').output == '✗ curse:molinari\n  package is pinned\n'
-    assert run('update curse:molinari').output == dedent(
+    assert run('update --retain-strategies curse:molinari').output == dedent(
         '''\
         ✓ curse:molinari
           updated 80000.57-Release to 90200.82-Release with new strategies: